separate type files
